--- frontend/jsconfig.json
@@ -1,10 +1,0 @@
-{
-  "compilerOptions": {
-    "jsx": "react",
-    "baseUrl": ".",
-    "paths": {
-      "@/*": ["./src/*"]
-    }
-  }
-}
-
