--- frontend/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "frontend",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@hookform/resolvers": "^5.0.1",
-    "autoprefixer": "^10.4.21",
-    "axios": "^1.8.4",
-    "next": "^15.3.0",
-    "postcss": "^8.5.3",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
-    "react-hook-form": "^7.55.0",
-    "react-hot-toast": "^2.5.2",
-    "socket.io-client": "^4.8.1",
-    "tailwindcss": "^4.1.4",
-    "zod": "^3.24.3"
-  },
-  "devDependencies": {
-    "@eslint/eslintrc": "^3",
-    "@types/node": "22.14.1",
-    "@types/react": "19.1.2",
-    "eslint": "^9",
-    "eslint-config-next": "15.3.0"
-  }
-}
