--- frontend/src/app/layout.js
@@ -1,20 +1,0 @@
-'use client';
-
-// Usa esta importación
-import { AuthProvider } from '../providers/AuthProvider';
-import { Inter } from 'next/font/google';
-import './globals.css';
-
-const inter = Inter({ subsets: ['latin'] });
-
-export default function RootLayout({ children }) {
-  return (
-    <html lang="es">
-      <body className={inter.className}suppressHydrationWarning={true}>
-        <AuthProvider>
-          {children}
-        </AuthProvider>
-      </body>
-    </html>
-  );
-}
